One tool-pattern matcher, and a consumer that proves it uses it
Closes #2118.

assay-metrics carried a byte-identical copy of assay-core's
matches_tool_pattern, with no test module, serving both policy.deny and
policy.allow. Per CLAUDE.md one rule gets one function and a parity test is
the fallback for when that is impossible; it was not impossible here.

The function moves to assay_common::tool_pattern, which both crates already
depend on. It meets the stated admission test for that crate: a mechanism
whose second implementation would silently mean something different. A
pattern language answers 'what does this policy permit', so two constructions
of it are two answers, and a policy file cannot show which one is being
applied to it. Nothing is std-gated -- every operation is a &str method and
nothing allocates, so the no_std build is unaffected.

The two had already begun to drift. #2117 deleted a dead 'pattern == "*"'
guard clause from the assay-core copy; the assay-metrics copy still had it.
That direction was harmless. The point is that nothing would have reported
the other one.

The divergence test stays in assay-core, which is the only crate depending on
both this matcher and the mandate glob in assay-evidence, so it is the only
place that comparison can be written.

Also here, and the reason this is not a pure move: breaking the shared
function left all 62 assay-metrics tests green. That crate's entire
tool-pattern path was uncovered, so unifying the code would have proved
nothing about the consumer. The new evaluator-level test pins the documented
interior-star behaviour where it actually bites -- 'deny: ["exec*sh"]' is a
literal name, does not match exec_bash, and permits the call. Its second half
denies the same call with an exact pattern, because asserting only that the
first call passes would hold if the policy never loaded at all.
