+++ crates/assay-core/src/mcp/policy/matcher.rs
@@ -1,137 +1,19 @@
-pub(in crate::mcp::policy) fn matches_tool_pattern(tool_name: &str, pattern: &str) -> bool {
-    if !pattern.contains('*') {
-        return tool_name == pattern;
-    }
-    let starts_star = pattern.starts_with('*');
-    let ends_star = pattern.ends_with('*');
-    match (starts_star, ends_star) {
-        (true, true) => {
-            let inner = pattern.trim_matches('*');
-            if inner.is_empty() {
-                true
-            } else {
-                tool_name.contains(inner)
-            }
-        }
-        (false, true) => {
-            let prefix = pattern.trim_end_matches('*');
-            !prefix.is_empty() && tool_name.starts_with(prefix)
-        }
-        (true, false) => {
-            let suffix = pattern.trim_start_matches('*');
-            !suffix.is_empty() && tool_name.ends_with(suffix)
-        }
-        (false, false) => tool_name == pattern,
-    }
-}
+//! The MCP policy tool-pattern matcher, and the one test that can only live here.
+//!
+//! The matcher itself moved to `assay_common::tool_pattern`, together with the tests that
+//! describe its behaviour, because `assay-metrics` held a byte-identical second copy of it and a
+//! pattern language is exactly the kind of mechanism a second implementation silently redefines.
+//!
+//! What stays is the divergence test. It compares this matcher against the mandate glob in
+//! `assay-evidence`, and `assay-core` is the only crate that depends on both, so it is the only
+//! place the comparison can be written at all.
+
+pub(in crate::mcp::policy) use assay_common::tool_pattern::matches_tool_pattern;
 
 #[cfg(test)]
 mod tests {
     use super::matches_tool_pattern;
     use crate::runtime::glob_matches_impl;
-
-    /// The five behaviours `docs/reference/config/policies.md` promises, pinned so the doc and the
-    /// matcher cannot drift apart silently. Nothing was checking that they agreed.
-    #[test]
-    fn the_documented_wildcard_forms_behave_as_documented() {
-        assert!(
-            matches_tool_pattern("anything_at_all", "*"),
-            "`*` matches all tools"
-        );
-        assert!(matches_tool_pattern("read_file", "read_*"), "prefix");
-        // Anchored, and the negative case has to contain the prefix somewhere other than the
-        // start or it proves nothing: `write_file` does not contain `read_` at all, so asserting
-        // on it holds even if prefix matching degraded to substring. Found by biting this test.
-        assert!(
-            !matches_tool_pattern("my_read_file", "read_*"),
-            "prefix is anchored at the start, not merely present"
-        );
-        assert!(matches_tool_pattern("read_file", "*_file"), "suffix");
-        assert!(
-            !matches_tool_pattern("_file_reader", "*_file"),
-            "suffix is anchored at the end, not merely present"
-        );
-        assert!(
-            matches_tool_pattern("web_search_api", "*search*"),
-            "substring"
-        );
-        assert!(
-            !matches_tool_pattern("web_fetch", "*search*"),
-            "substring is real"
-        );
-        assert!(matches_tool_pattern("exec", "exec"), "no star is exact");
-        assert!(
-            !matches_tool_pattern("exec_shell", "exec"),
-            "no star is not a prefix"
-        );
-    }
-
-    /// A `*` that is neither the first nor the last character is a literal, and the doc did not
-    /// say so. `read_*_file` does not match `read_config_file`; it matches only a tool literally
-    /// named `read_*_file`, because the pattern falls through to the exact-match arm with the
-    /// asterisk still in it.
-    ///
-    /// The consequence runs the dangerous way. In `allow`, such a pattern matches nothing and the
-    /// tool is refused -- fail-closed, visible. In `deny`, it matches nothing and the tool is
-    /// **permitted**, silently, by a line whose author believed it was blocking something.
-    ///
-    /// It is not only the exact-match arm. A pattern that is a wildcard at both ends carries its
-    /// interior stars into the substring it searches for, so `*a*b*` looks for the three
-    /// characters `a*b` and finds them only in a tool name that really contains an asterisk.
-    #[test]
-    fn a_star_in_the_middle_is_a_literal_not_a_wildcard() {
-        assert!(
-            !matches_tool_pattern("read_config_file", "read_*_file"),
-            "a middle star does not expand"
-        );
-        assert!(
-            matches_tool_pattern("read_*_file", "read_*_file"),
-            "it is matched literally, asterisk included"
-        );
-        assert!(
-            !matches_tool_pattern("abc", "a*b*c"),
-            "multiple interior stars are literal too"
-        );
-        assert!(
-            !matches_tool_pattern("axbc", "*a*b*"),
-            "an interior star inside the substring form is literal as well"
-        );
-        assert!(
-            matches_tool_pattern("a*b", "*a*b*"),
-            "which is to say the substring searched for is `a*b`, asterisk and all"
-        );
-    }
-
-    /// Only-stars and the empty pattern, so the corners are stated rather than discovered.
-    ///
-    /// The first version of this test asserted `!matches_tool_pattern("anything", "")` under the
-    /// message "an empty pattern matches nothing". The assertion held and the message was false:
-    /// an empty pattern has no star, so it takes the exact-match arm and matches the empty tool
-    /// name. A corner stated wrongly is worse than a corner left undocumented, and this test
-    /// exists to state corners.
-    #[test]
-    fn degenerate_patterns_have_stated_answers() {
-        assert!(
-            matches_tool_pattern("anything", "**"),
-            "leading+trailing star is the substring form with an empty inner"
-        );
-        assert!(
-            matches_tool_pattern("anything", "***"),
-            "same, whatever the count"
-        );
-        assert!(
-            !matches_tool_pattern("anything", ""),
-            "an empty pattern is an exact match against the empty name, so it matches no real tool"
-        );
-        assert!(
-            matches_tool_pattern("", ""),
-            "and it does match the empty name, which is what makes it exact rather than inert"
-        );
-        assert!(
-            matches_tool_pattern("", "*"),
-            "`*` matches the empty tool name too"
-        );
-    }
 
     /// Two matchers, one character, several meanings — and the differences are deliberate.
     ///
@@ -200,26 +82,5 @@ mod tests {
         // `**` means nothing special to the policy matcher: leading and trailing star is the
         // substring form, so it matches for a different reason than a reader might assume.
         assert!(matches_tool_pattern("fs.read_file", "**"));
-    }
-
-    /// An unbounded `*` over-blocks in `deny` and over-permits in `allow`. Only the second is
-    /// dangerous, and the asymmetry is invisible in a policy file, which is why the reference
-    /// states it. Azure RBAC measured the cost of the permissive direction: about half of the
-    /// 15,481 catalogued actions reach across Resource Providers under non-obvious wildcards
-    /// (arXiv 2506.10755v3), and the recommendation is explicit enumeration rather than a
-    /// cleverer pattern language.
-    ///
-    /// Note what this test cannot show. `*` reaches the same answer through the empty-inner
-    /// substring arm, so no branch is uniquely its own -- an earlier version of this file carried
-    /// a `pattern == "*"` guard clause above the match, and deleting it changed nothing. A test
-    /// named for a pattern is not a test of a branch.
-    #[test]
-    fn a_star_in_allow_admits_every_tool_including_ones_not_yet_written() {
-        for tool in ["read_file", "exec", "shell", "a.tool.added.next.week"] {
-            assert!(
-                matches_tool_pattern(tool, "*"),
-                "`allow: ['*']` admits {tool}"
-            );
-        }
     }
 }
