+++ crates/assay-metrics/src/args_valid_next/matcher.rs
@@ -1,29 +1,13 @@
-pub(super) fn matches_tool_pattern(tool_name: &str, pattern: &str) -> bool {
-    if pattern == "*" {
-        return true;
-    }
-    if !pattern.contains('*') {
-        return tool_name == pattern;
-    }
-    let starts_star = pattern.starts_with('*');
-    let ends_star = pattern.ends_with('*');
-    match (starts_star, ends_star) {
-        (true, true) => {
-            let inner = pattern.trim_matches('*');
-            if inner.is_empty() {
-                true
-            } else {
-                tool_name.contains(inner)
-            }
-        }
-        (false, true) => {
-            let prefix = pattern.trim_end_matches('*');
-            !prefix.is_empty() && tool_name.starts_with(prefix)
-        }
-        (true, false) => {
-            let suffix = pattern.trim_start_matches('*');
-            !suffix.is_empty() && tool_name.ends_with(suffix)
-        }
-        (false, false) => tool_name == pattern,
-    }
-}
+//! The tool-pattern matcher this module used to define now lives in `assay_common`.
+//!
+//! It was a byte-identical copy of `assay-core`'s, with no tests of its own, serving both
+//! `policy.deny` and `policy.allow` in `evaluator.rs`. The two had already begun to drift: when
+//! `assay-core` dropped a dead `pattern == "*"` guard clause in #2117, this copy kept it. That is
+//! the harmless direction of drift, and the point is that nothing would have reported the other
+//! one — the interior-star rule (`read_*_file` is a literal, so a `deny` entry shaped like a glob
+//! blocks nothing) held here too, unstated and unpinned.
+//!
+//! Per `CLAUDE.md`, one rule gets one function; a parity test is the fallback for when that is
+//! impossible, and here it was not impossible.
+
+pub(super) use assay_common::tool_pattern::matches_tool_pattern;
